Skeletons for the event created
Using proxy, the skeletons for the event model is designed. If you want to add attributes that all can share, you can add it on the Event class, but if they have unique attributes and fields, you can add it to their specific name + More classes. Eg. If you want to add unique attributes for the Hackathon type, you can add it on HackathonsMore class.